feat(#279): add frequently asked questions
Q: "Since Mycelium is the network structure, and utility is the core flow within it, using Trehalose as a naturally occurring sugar, is that the idea?"

A: "I really like the analogy. Trehalose is used as a common nutrient interface to exchange value between all the plants connected through the mycelium. It feels quite similar how our DAO mycel is exchanging value."