feat(#279): add frequently asked questions
Q: "How is the stability ensured? Is it pegged to fiat, algorithmic, or backed by a basket of assets?"

A: "My idea is to couple global inflation somehow. We could do this by using all the bills we receive for our global deployment. By that we can calculate the price per compute unit at every given time. This price defines the value of TREH. By this, the user always receives predictable price for platform usage."