feat(#279): add frequently asked questions
Q: "Would it make sense to further uncouple the token from the DAO creation process? i.e. we send you TreehaTokens (airdrop) and you do the rest?"

A: "No, absolutely not. The intention is to use the SCW as a bot protection layer. It doesn't allow anything other than sending tokens back to treasury based on your usage patterns. It's a non-custodial smart wallet controlled by the protocol and connected to the individual space."